+++ zoom/apps.py
@@ -683,19 +683,25 @@ def main_menu_items(request):
     def style(app):
         return app.name == current_app_name and 'class="active" ' or ''
 
+    def li(app_name, body):
+        return f'<li class="{app_name}-menu-item">{body}</li>'
+
     default_app_name = get_default_app_name(request.site, request.user)
     current_app_name = request.route and request.route[0] or default_app_name
 
-    return html.li([
-        app.name == default_app_name
-        and '<a {}href="<dz:site_url>/">{}</a>'.format(
-            style(app), app.title
-        )
-        or '<a {}href="<dz:site_url>{}">{}</a>'.format(
-            style(app), app.url, app.title
+    return ''.join(
+        li(
+            app.name,
+            app.name == default_app_name
+            and '<a {}href="<dz:site_url>/">{}</a>'.format(
+                style(app), app.title
+            )
+            or '<a {}href="<dz:site_url>{}">{}</a>'.format(
+                style(app), app.url, app.title
+            )
         )
         for app in get_main_apps(request) if app.visible
-    ])
+    )
 
 
 def main_menu(request):