+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add app badges
 - wrap alerts in div
 - change title bar to use css grid
+- adjust PyMySQL requirement version
 
 ## [6.22.0] - 2023-11-20
 - add support for custom MySQL port